+++ src/plugins/intel_gpu/include/intel_gpu/plugin/graph.hpp
@@ -59,7 +59,7 @@ public:
     cldnn::engine& get_engine() const { return m_context->get_engine(); }
     const ExecutionConfig& get_config() const { return m_config; }
 
-    int GetMaxDynamicBatchSize() const { return static_cast<int>(m_config.get_property(ov::intel_gpu::max_dynamic_batch));}
+    size_t GetMaxDynamicBatchSize() const { return m_config.get_property(ov::intel_gpu::max_dynamic_batch);}
     const std::map<std::string, cldnn::layout>& GetInputLayouts() const { return m_program->GetInputLayouts(); }
     const InferenceEngine::InputsDataMap GetNetworkInputs() const { return m_program->GetNetworkInputs(); }
     const InferenceEngine::OutputsDataMap GetNetworkOutputs() const { return m_program->GetNetworkOutputs(); }
